requirements.txt
ruamel.yaml==0.17.21
jsonschema==4.4.0
SPARQLWrapper==2.0.0
ordered-set==4.1.0
# cap-pipeline-config-utils==0.0.1